ExtractAccessKey now stores position
git-svn-id: svn://ultimatepp.org/upp/trunk@997 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlCore/CtrlCore.h
@@ -1147,7 +1147,7 @@ public:
 	static bool  IterateFocusForward(Ctrl *ctrl, Ctrl *top, bool noframe = false, bool init = false, bool all = false);
 	static bool  IterateFocusBackward(Ctrl *ctrl, Ctrl *top, bool noframe = false, bool all = false);
 
-	static dword AccessKeyBit(byte accesskey);
+	static dword AccessKeyBit(int accesskey);
 	dword   GetAccessKeysDeep() const;
 	void    DistributeAccessKeys();
 	bool    VisibleAccessKeys();
