+++ uppsrc/RichText/RichText.h
@@ -382,7 +382,7 @@ public:
 
 String AsQTF(const RichObject& obj);
 
-RichText ParseQTF(const char *qtf, byte accesskey = 0);
+RichText ParseQTF(const char *qtf, int accesskey = 0);
 
 RichText AsRichText(const wchar *s, const RichPara::Format& f = RichPara::Format());
